requirements.txt
httpx
python-dotenv
aiosqlite
anthropic
openai
fastapi
uvicorn[standard]
jinja2
python-multipart
